13-newsletter-agent/requirements.txt
google-adk[database]==0.3.0
google-generativeai==0.8.5
python-dotenv==1.1.0
feedparser>=6.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
sqlalchemy>=2.0.0
